### 2023-07-22 #5680 @editorialbot
[PRE REVIEW]: epiworldR: Fast Agent-Based Epi Models
**Submitting author:** @derekmeyer37 (<a href="http://orcid.org/0009-0005-1350-6988">Derek  Meyer</a>)
**Repository:** https://github.com/UofUEpiBio/epiworldR
**Branch with paper.md** (empty if default branch):
**Version:** 0.0-2
