+++ man/ModelSIS.Rd
@@ -68,6 +68,7 @@ Other Models:
 \code{\link{ModelSEIRCONN}()},
 \code{\link{ModelSEIR}()},
 \code{\link{ModelSIRCONN}()},
+\code{\link{ModelSIRD}()},
 \code{\link{ModelSIRLogit}()},
 \code{\link{ModelSIR}()},
 \code{\link{ModelSURV}()},
