+++ man/ModelSEIRCONN.Rd
@@ -88,6 +88,7 @@ Other Models:
 \code{\link{ModelDiffNet}()},
 \code{\link{ModelSEIR}()},
 \code{\link{ModelSIRCONN}()},
+\code{\link{ModelSIRD}()},
 \code{\link{ModelSIRLogit}()},
 \code{\link{ModelSIR}()},
 \code{\link{ModelSIS}()},
