### 2023-07-26 a5e032c
Adding SIRDCONN
