### 2023-07-26 98f5f33
Adding SEIRDCONN
