### 2023-07-27 4796f06
Adding ModelSIRDCONN
