+++ man/ModelSIS.Rd
@@ -66,11 +66,14 @@ epiworld-methods
 Other Models: 
 \code{\link{ModelDiffNet}()},
 \code{\link{ModelSEIRCONN}()},
+\code{\link{ModelSEIRDCONN}()},
 \code{\link{ModelSEIR}()},
 \code{\link{ModelSIRCONN}()},
 \code{\link{ModelSIRD}()},
 \code{\link{ModelSIRLogit}()},
 \code{\link{ModelSIR}()},
+\code{\link{ModelSISD}()},
+\code{\link{ModelSURVD}()},
 \code{\link{ModelSURV}()},
 \code{\link{epiworld-data}}
 }

[…]
