+++ R/ModelSIRDCONN.R
@@ -1,4 +1,4 @@
-#' Susceptible Infected Removed model (SIR connected)
+#' Susceptible Infected Removed Deceased model (SIRD connected)
 #' @param name String. Name of the virus
 #' @param prevalence Double. Initial proportion of individuals with the virus.
 #' @param contact_rate Numeric scalar. Average number of contacts per step.
@@ -6,48 +6,50 @@
 #' transmission.
 #' @param recovery_rate Numeric scalar between 0 and 1. Probability of recovery.
 #' @param death_rate Numeric scalar between 0 and 1. Probability of death.
-#' @param x Object of class SIRCONN. 
+#' @param x Object of class SIRDCONN. 
 #' @param ... Currently ignore.
 #' @param n Number of individuals in the population.
 #' @export
 #' @family Models
 #' @returns
 #' - The `ModelSIRDCONN`function returns a model of class [epiworld_model].
-#' @aliases epiworld_sirconn
+#' @aliases epiworld_sirdconn
 #' 
 #' @examples 
-#' model_sirconn <- ModelSIRCONN(
+#' model_sirdconn <- ModelSIRDCONN(
 #'   name                = "COVID-19",
 #'   n                   = 10000,
 #'   prevalence          = 0.01,
 #'   contact_rate        = 5,
 #'   transmission_rate   = 0.4,
-#'   recovery_rate       = 0.95
+#'   recovery_rate       = 0.95,
+#'   death_rate          = 0.01
 #' )
 #'   
 #' # Running and printing
-#' run(model_sirconn, ndays = 100, seed = 1912)
-#' model_sirconn
+#' run(model_sirdconn, ndays = 100, seed = 1912)
+#' model_sirdconn
 #' 
-#' plot(model_sirconn,  main = "SIRCONN Model")
+#' plot(model_sirdconn,  main = "SIRDCONN Model")
 #' @seealso epiworld-methods
-ModelSIRCONN <- function(
-    name, n, prevalence, contact_rate, transmission_rate, recovery_rate
+ModelSIRDCONN <- function(
+    name, n, prevalence, contact_rate, transmission_rate, recovery_rate,
+    death_rate
 ) {
   
   structure(
-    ModelSIRCONN_cpp(name, n, prevalence, contact_rate, 
-                     transmission_rate, recovery_rate),
-    class = c("epiworld_sirconn", "epiworld_model")
+    ModelSIRDCONN_cpp(name, n, prevalence, contact_rate, 
+                     transmission_rate, recovery_rate, death_rate),
+    class = c("epiworld_sirdconn", "epiworld_model")
   )
   
 }
 
-#' @rdname ModelSIRCONN
+#' @rdname ModelSIRDCONN
 #' @export
-#' @returns The `plot` function returns a plot of the SIRCONN model of class 
+#' @returns The `plot` function returns a plot of the SIRDCONN model of class 
 #' [epiworld_model].
 #' @param main Title of the plot
-plot.epiworld_sirconn <- function(x, main = get_name(x), ...) { # col = NULL
+plot.epiworld_sirdconn <- function(x, main = get_name(x), ...) { # col = NULL
  plot_epi(x, main = main, ...)
 }
