+++ man/ModelSIRCONN.Rd
@@ -1,22 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ModelSIRCONN.R, R/ModelSIRDCONN.R
+% Please edit documentation in R/ModelSIRCONN.R
 \name{ModelSIRCONN}
 \alias{ModelSIRCONN}
 \alias{epiworld_sirconn}
 \alias{plot.epiworld_sirconn}
 \title{Susceptible Infected Removed model (SIR connected)}
 \usage{
-ModelSIRCONN(
-  name,
-  n,
-  prevalence,
-  contact_rate,
-  transmission_rate,
-  recovery_rate
-)
-
-\method{plot}{epiworld_sirconn}(x, main = get_name(x), ...)
-
 ModelSIRCONN(
   name,
   n,
@@ -47,8 +36,6 @@ transmission.}
 \item{main}{Title of the plot}
 
 \item{...}{Currently ignore.}
-
-\item{death_rate}{Numeric scalar between 0 and 1. Probability of death.}
 }
 \value{
 \itemize{
@@ -57,34 +44,11 @@ transmission.}
 
 The \code{plot} function returns a plot of the SIRCONN model of class
 \link{epiworld_model}.
-
-\itemize{
-\item The \code{ModelSIRDCONN}function returns a model of class \link{epiworld_model}.
-}
-
-The \code{plot} function returns a plot of the SIRCONN model of class
-\link{epiworld_model}.
 }
 \description{
 Susceptible Infected Removed model (SIR connected)
-
-Susceptible Infected Removed model (SIR connected)
 }
 \examples{
-model_sirconn <- ModelSIRCONN(
-  name                = "COVID-19",
-  n                   = 10000,
-  prevalence          = 0.01,
-  contact_rate        = 5,
-  transmission_rate   = 0.4,
-  recovery_rate       = 0.95
-)
-  
-# Running and printing
-run(model_sirconn, ndays = 100, seed = 1912)
-model_sirconn
-
-plot(model_sirconn,  main = "SIRCONN Model")
 model_sirconn <- ModelSIRCONN(
   name                = "COVID-19",
   n                   = 10000,
@@ -103,27 +67,12 @@ plot(model_sirconn,  main = "SIRCONN Model")
 \seealso{
 epiworld-methods
 
-epiworld-methods
-
 Other Models: 
 \code{\link{ModelDiffNet}()},
 \code{\link{ModelSEIRCONN}()},
 \code{\link{ModelSEIRDCONN}()},
 \code{\link{ModelSEIR}()},
-\code{\link{ModelSIRD}()},
-\code{\link{ModelSIRLogit}()},
-\code{\link{ModelSIR}()},
-\code{\link{ModelSISD}()},
-\code{\link{ModelSIS}()},
-\code{\link{ModelSURVD}()},
-\code{\link{ModelSURV}()},
-\code{\link{epiworld-data}}
-
-Other Models: 
-\code{\link{ModelDiffNet}()},
-\code{\link{ModelSEIRCONN}()},
-\code{\link{ModelSEIRDCONN}()},
-\code{\link{ModelSEIR}()},
+\code{\link{ModelSIRDCONN}()},
 \code{\link{ModelSIRD}()},
 \code{\link{ModelSIRLogit}()},
 \code{\link{ModelSIR}()},

[…]
