### 2023-08-10 8848202
Fixing survd deletion
