+++ inst/include/epiworld.hpp
@@ -17330,418 +17330,6 @@ inline void ModelSEIRD<TSeq>::contact(Model<TSeq> * m)
 /*//////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
 
- Start of -include/epiworld//models/surveillanced.hpp-
-
-////////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////*/
-
-
-#ifndef EPIWORLD_MODELS_SURVEILLANCED_HPP
-#define EPIWORLD_MODELS_SURVEILLANCED_HPP
-
-template<typename TSeq = EPI_DEFAULT_TSEQ>
-class ModelSURVD : public epiworld::Model<TSeq> {
-
-private:
-    // state
-    static const int SUSCEPTIBLE           = 0;
-    static const int LATENT                = 1;
-    static const int SYMPTOMATIC           = 2;
-    static const int SYMPTOMATIC_ISOLATED  = 3; // sampled and discovered
-    static const int ASYMPTOMATIC          = 4;
-    static const int ASYMPTOMATIC_ISOLATED = 5;
-    static const int RECOVERED             = 6;
-    static const int REMOVED               = 7;
-
-public:
-
-    /**
-     * @name Construct a new ModelSURVD object
-     * 
-     * The ModelSURVD class simulates a survaillence model where agents can be
-     * isolated, even if asyptomatic.
-     * 
-     * @param vname String. Name of the virus
-     * @param prevalence Integer. Number of initial cases of the virus.
-     * @param efficacy_vax Double. Efficacy of the vaccine (1 - P(acquire the disease)).
-     * @param latent_period Double. Shape parameter of a `Gamma(latent_period, 1)`
-     *   distribution. This coincides with the expected number of latent days.
-     * @param infect_period Double. Shape parameter of a `Gamma(infected_period, 1)`
-     *   distribution. This coincides with the expected number of infectious days.
-     * @param prob_symptoms Double. Probability of generating symptoms.
-     * @param prop_vaccinated Double. Probability of vaccination. Coincides with
-     *   the initial prevalence of vaccinated individuals.
-     * @param prop_vax_redux_transm Double. Factor by which the vaccine reduces
-     *   transmissibility.
-     * @param prop_vax_redux_infect Double. Factor by which the vaccine reduces
-     *   the chances of becoming infected.
-     * @param surveillance_prob Double. Probability of testing an agent.
-     * @param prob_transmission Double. Raw transmission probability.
-     * @param prob_death Double. Raw probability of death for symptomatic individuals.
-     * @param prob_noreinfect Double. Probability of no re-infection.
-     * 
-     * @details
-     * This model features the following states:
-     * 
-     * - Susceptible
-     * - Latent
-     * - Symptomatic
-     * - Symptomatic isolated
-     * - Asymptomatic
-     * - Asymptomatic isolated
-     * - Recovered
-     * - Removed  
-     * - Deceased  
-     * 
-     * @returns An object of class `epiworld_survd`
-     * 
-     */
-    ///@{
-    ModelSURVD() {};
-
-    ModelSURVD(
-        ModelSURVD<TSeq> & model,
-        std::string vname,
-        epiworld_fast_uint prevalence               = 50,
-        epiworld_double efficacy_vax          = 0.9,
-        epiworld_double latent_period         = 3u,
-        epiworld_double infect_period         = 6u,
-        epiworld_double prob_symptoms         = 0.6,
[…]
