+++ man/epiworld-data.Rd
@@ -246,7 +246,6 @@ Other Models:
 \code{\link{ModelSIR}()},
 \code{\link{ModelSISD}()},
 \code{\link{ModelSIS}()},
-\code{\link{ModelSURVD}()},
 \code{\link{ModelSURV}()}
 }
 \concept{Models}
