+++ man/ModelSISD.Rd
@@ -1,0 +1,87 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ModelSISD.R
+\name{ModelSISD}
+\alias{ModelSISD}
+\alias{epiworld_sisd}
+\alias{plot.epiworld_sisd}
+\title{SISD model}
+\usage{
+ModelSISD(name, prevalence, transmission_rate, recovery_rate, death_rate)
+
+\method{plot}{epiworld_sisd}(x, main = get_name(x), ...)
+}
+\arguments{
+\item{name}{String. Name of the virus.}
+
+\item{prevalence}{Double. Initial proportion of individuals with the virus.}
+
+\item{transmission_rate}{Numeric scalar between 0 and 1. Virus's rate of
+infection.}
+
+\item{recovery_rate}{Numeric scalar between 0 and 1. Rate of recovery from virus.}
+
+\item{death_rate}{Numeric scalar between 0 and 1. Rate of death from virus.}
+
+\item{x}{Object of class SISD.}
+
+\item{main}{Title of the plot.}
+
+\item{...}{Currently ignore.}
+}
+\value{
+\itemize{
+\item The \code{ModelSISD} function returns a model of class \link{epiworld_model}.
+}
+
+\itemize{
+\item The \code{plot} function returns a plot of the SISD model of class
+\link{epiworld_model}.
+}
+}
+\description{
+Susceptible-Infected-Susceptible-Deceased model (SISD) (\href{https://en.wikipedia.org/w/index.php?title=Compartmental_models_in_epidemiology&oldid=1155757336#The_SIS_model}{wiki})
+}
+\examples{
+model_sisd <- ModelSISD(
+               name = "COVID-19", 
+               prevalence = 0.01, 
+               transmission_rate = 0.9, 
+               recovery_rate = 0.1, 
+               death_rate = 0.01
+             )
+
+# Adding a small world population
+agents_smallworld(
+  model_sisd,
+  n = 1000,
+  k = 5,
+  d = FALSE,
+  p = .01
+  )
+  
+# Running and printing
+run(model_sisd, ndays = 100, seed = 1912)
+model_sisd
+
+# Plotting
+plot(model_sisd, main = "SISD Model")
+
+}
+\seealso{
+epiworld-methods
+
+Other Models: 
+\code{\link{ModelDiffNet}()},
+\code{\link{ModelSEIRCONN}()},
+\code{\link{ModelSEIRDCONN}()},
+\code{\link{ModelSEIR}()},
+\code{\link{ModelSIRCONN}()},
+\code{\link{ModelSIRDCONN}()},
+\code{\link{ModelSIRD}()},
+\code{\link{ModelSIRLogit}()},
+\code{\link{ModelSIR}()},
+\code{\link{ModelSIS}()},
+\code{\link{ModelSURV}()},
+\code{\link{epiworld-data}}
+}
+\concept{Models}
