+++ NEWS.md
@@ -1,3 +1,9 @@
+# epiworldR 0.0-3
+
+* Added the following models: `ModelSEIRD`, `ModelSEIRDCONN`, `ModelSIRD`, `ModelSIRDCONN`, `ModelSISD`, and `ModelSURVD`.
+
+* Fixed a bug reported on issue [6](https://github.com/UofUEpiBio/epiworldR/issues/6).
+
 # epiworldR 0.0-2
 
 * Added a `NEWS.md` file to track changes to the package.

[…]
