+++ inst/include/epiworld.hpp
@@ -15482,7 +15482,7 @@ inline ModelSIR<TSeq>::ModelSIR(
  * @param vname std::string Name of the virus
  * @param prevalence epiworld_double Initial prevalence the immune system
  * @param transmission_rate epiworld_double Transmission rate of the virus
- * @param avg_incubation_days epiworld_double Average incubation days of the virus
+ * @param avg_incubation_days epiworld_double Average incubation days of the virus.
  * @param recovery_rate epiworld_double Recovery rate of the virus.
  */
 template<typename TSeq = int>
@@ -16974,26 +16974,24 @@ inline ModelSISD<TSeq>::ModelSISD(
 
 /**
  * @brief Template for a Susceptible-Exposed-Infected-Removed-Deceased (SEIRD) model
-*/
+ * 
+ * @param model A Model<TSeq> object where to set up the SIR.
+ * @param vname std::string Name of the virus
+ * @param prevalence epiworld_double Initial prevalence the immune system
+ * @param transmission_rate epiworld_double Transmission rate of the virus
+ * @param avg_incubation_days epiworld_double Average incubation days of the virus
+ * @param recovery_rate epiworld_double Recovery rate of the virus.
+ * @param death_rate epiworld_double Death rate of the virus.
+ */
 template<typename TSeq = int>
 class ModelSEIRD : public epiworld::Model<TSeq>
 {
-
-protected:
-
-    enum S {
-        Susceptible,
-        Exposed,
-        Infected,
-        Hospitalized,
-        Recovered,
-        Deceased
-    };
-    
-    static void update_exposed(epiworld::Agent<TSeq> * p, epiworld::Model<TSeq> * m);
-    static void update_infected(epiworld::Agent<TSeq> * p, epiworld::Model<TSeq> * m);
-    static void update_hospitalized(epiworld::Agent<TSeq> * p, epiworld::Model<TSeq> * m);
-    static void contact(Model<TSeq> * m);
+private:
+    static const int SUSCEPTIBLE = 0;
+    static const int EXPOSED     = 1;
+    static const int INFECTED    = 2;
+    static const int REMOVED     = 3;
+    static const int DECEASED    = 4;
 
 public:
 
@@ -17001,116 +16999,134 @@ public:
 
     ModelSEIRD(
         ModelSEIRD<TSeq> & model,
-        std::string vname
+        std::string vname,
+        epiworld_double prevalence,
+        epiworld_double transmission_rate,
+        epiworld_double avg_incubation_days,
+        epiworld_double recovery_rate,
+        epiworld_double death_rate
     );
 
     ModelSEIRD(
         std::string vname,
         epiworld_double prevalence,
-        epiworld_double incu_shape,
-        epiworld_double incu_rate,
-        epiworld_double infe_shape,
-        epiworld_double infe_rate,
-        epiworld_double hospitalization_rate,
-        epiworld_double hospitalization_recover_rate,
-        epiworld_double hospitalization_decease_rate,
         epiworld_double transmission_rate,
-        epiworld_double transmission_entity_rate,
-        size_t n_entities,
-        size_t n_interactions
+        epiworld_double avg_incubation_days,
+        epiworld_double recovery_rate,
+        epiworld_double death_rate
     );
-
-    ModelSEIRD(
-        std::string fn,
-        std::string vname
-        );
     
+    epiworld::UpdateFun<TSeq> update_exposed_seir = [](
+        epiworld::Agent<TSeq> * p,
+        epiworld::Model<TSeq> * m
+    ) -> void {
+
+        // Getting the virus
+        auto v = p->get_virus(0);
+
+        // Does the agent become infected?
[…]
