+++ man/ModelSIRD.Rd
@@ -73,6 +73,7 @@ Other Models:
 \code{\link{ModelDiffNet}()},
 \code{\link{ModelSEIRCONN}()},
 \code{\link{ModelSEIRDCONN}()},
+\code{\link{ModelSEIRD}()},
 \code{\link{ModelSEIR}()},
 \code{\link{ModelSIRCONN}()},
 \code{\link{ModelSIRDCONN}()},
