+++ inst/include/epiworld.hpp
@@ -16987,106 +16987,114 @@ template<typename TSeq = int>
 class ModelSEIRD : public epiworld::Model<TSeq>
 {
 private:
-    static const int SUSCEPTIBLE = 0;
-    static const int EXPOSED     = 1;
-    static const int INFECTED    = 2;
-    static const int REMOVED     = 3;
-    static const int DECEASED    = 4;
-
+  static const int SUSCEPTIBLE = 0;
+  static const int EXPOSED     = 1;
+  static const int INFECTED    = 2;
+  static const int REMOVED     = 3;
+  static const int DECEASED    = 4;
+  
 public:
-
-    ModelSEIRD() {};
-
-    ModelSEIRD(
-        ModelSEIRD<TSeq> & model,
-        std::string vname,
-        epiworld_double prevalence,
-        epiworld_double transmission_rate,
-        epiworld_double avg_incubation_days,
-        epiworld_double recovery_rate,
-        epiworld_double death_rate
-    );
-
-    ModelSEIRD(
-        std::string vname,
-        epiworld_double prevalence,
-        epiworld_double transmission_rate,
-        epiworld_double avg_incubation_days,
-        epiworld_double recovery_rate,
-        epiworld_double death_rate
-    );
+  
+  ModelSEIRD() {};
+  
+  ModelSEIRD(
+    ModelSEIRD<TSeq> & model,
+    std::string vname,
+    epiworld_double prevalence,
+    epiworld_double transmission_rate,
+    epiworld_double avg_incubation_days,
+    epiworld_double recovery_rate,
+    epiworld_double death_rate
+  );
+  
+  ModelSEIRD(
+    std::string vname,
+    epiworld_double prevalence,
+    epiworld_double transmission_rate,
+    epiworld_double avg_incubation_days,
+    epiworld_double recovery_rate,
+    epiworld_double death_rate
+  );
+  
+  epiworld::UpdateFun<TSeq> update_exposed_seir = [](
+    epiworld::Agent<TSeq> * p,
+    epiworld::Model<TSeq> * m
+  ) -> void {
     
-    epiworld::UpdateFun<TSeq> update_exposed_seir = [](
-        epiworld::Agent<TSeq> * p,
-        epiworld::Model<TSeq> * m
-    ) -> void {
-
-        // Getting the virus
-        auto v = p->get_virus(0);
-
-        // Does the agent become infected?
-        if (m->runif() < 1.0/(v->get_incubation(m)))
-            p->change_state(m, ModelSEIRD<TSeq>::INFECTED);
-
-        return;    
-    };
+    // Getting the virus
+    auto v = p->get_virus(0);
+    
+    // Does the agent become infected?
+    if (m->runif() < 1.0/(v->get_incubation(m)))
+      p->change_state(m, ModelSEIRD<TSeq>::INFECTED);
+    
+    return;    
+  };
+  
+  
+  epiworld::UpdateFun<TSeq> update_infected = [](
+    epiworld::Agent<TSeq> * p, epiworld::Model<TSeq> * m
+  ) -> void {
+    
+    auto state = p->get_state();
+    
+    if (state == ModelSEIRD<TSeq>::INFECTED)
+    {
       
-
-      epiworld::UpdateFun<TSeq> update_infected = [](
-        epiworld::Agent<TSeq> * p, epiworld::Model<TSeq> * m
-      ) -> void {
+      
+      // Odd: Die, Even: Recover
+      epiworld_fast_uint n_events = 0u;
+      for (const auto & v : p->get_viruses())
+      {
         
-        auto state = p->get_state();
+        // Recover
+        m->array_double_tmp[n_events++] = 
[…]
