### 2023-08-24 #5781 @editorialbot
[REVIEW]: epiworldR: Fast Agent-Based Epi Models
**Submitting author:** @derekmeyer37 (<a href="http://orcid.org/0009-0005-1350-6988">Derek  Meyer</a>)
**Repository:** https://github.com/UofUEpiBio/epiworldR
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3-0
