### 2023-08-24 #5781 @carinogurjao
Hello @spholmes ,
Here are my comments for the authors:

- The software paper needs a clear "References section"
- The installation with devtools `devtools::install_github("UofUEpiBio/epiworldR") `throws the following error. The installation with `install.packages("epiworldR") `works fine though.
```
Warning: S3 method ‘plot.epiworld_seird’ was declared in NAMESPACE but not found
Error: package or namespace load failed for ‘epiworldR’ in namespaceExport(ns, exports):
 undefined exports: ModelSEIRD
Error: loading failed
Execution halted
ERROR: loading failed
```
- netplot should be included as a dependency (for the Transmission network analysis)
- I don't see Community guidelines
