### 2023-08-28 aa0ada2
Updating virus methods
