+++ NAMESPACE
@@ -6,8 +6,12 @@ S3method(add_tool_n,epiworld_model)
 S3method(add_virus,epiworld_model)
 S3method(add_virus,epiworld_seir)
 S3method(add_virus,epiworld_seirconn)
+S3method(add_virus,epiworld_seird)
+S3method(add_virus,epiworld_seirdconn)
 S3method(add_virus,epiworld_sir)
 S3method(add_virus,epiworld_sirconn)
+S3method(add_virus,epiworld_sird)
+S3method(add_virus,epiworld_sirdconn)
 S3method(add_virus_n,epiworld_model)
 S3method(add_virus_n,epiworld_seir)
 S3method(add_virus_n,epiworld_seirconn)
