### 2023-08-29 a24b042
Fixing GHA (plot run_multipl)
