+++ README.md
@@ -1,6 +1,14 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
+## Versions
+
+The virtual INSNA Sunbelt 2023 session can be found here:
+<https://github.com/UofUEpiBio/epiworldR-workshop/tree/sunbelt2023-virtual>
+
+The in-person INSNA Sunbelt 2023 session can be found here:
+<https://github.com/UofUEpiBio/epiworldR-workshop/tree/sunbetl2023-inperson>
+
 # epiworldR
 
 <!-- badges: start -->
@@ -117,7 +125,7 @@ summary(sir)
 #> Number of entities  : 0
 #> Days (duration)     : 50 (of 50)
 #> Number of viruses   : 1
-#> Last run elapsed t  : 194.00ms
+#> Last run elapsed t  : 193.00ms
 #> Last run speed      : 25.77 million agents x day / second
 #> Rewiring            : off
 #> 
@@ -409,17 +417,42 @@ head(ans$reproductive)
 plot(ans$reproductive)
 ```
 
-<img src="man/figures/README-unnamed-chunk-6-1.png" width="100%" />
+<img src="man/figures/README-unnamed-chunk-6-1.png" width="100%" /> \#
+Existing Alternatives
 
-# Other ABM R packages
-
-Several alternatives to `epiworldR` exist and provide researchers with a range of options, each with its own unique features and strengths, enabling the exploration and analysis of infectious disease dynamics through agent-based modeling. Below is a manually curated table of existing alternatives including ABM [@ABM], abmR [@abmR], cystiSim [@cystiSim], villager [@villager], and RNetLogo [@RNetLogo].
+Several alternatives to `epiworldR` exist and provide researchers with a
+range of options, each with its own unique features and strengths,
+enabling the exploration and analysis of infectious disease dynamics
+through agent-based modeling. Below is a manually curated table of
+existing alternatives including ABM \[@ABM\], abmR \[@abmR\], cystiSim
+\[@cystiSim\], villager \[@villager\], and RNetLogo \[@RNetLogo\].
 
 | Package                                                                     | Multiple Viruses | Multiple Tools | Multiple Runs | Global Actions | Built-In Epi Models | Dependencies                                                                                             | Activity                                                                                                               |
-|:--------|:--------|:--------|:--------|:--------|---------|:--------|:--------|
+|:----------------------------------------------------------------------------|:-----------------|:---------------|:--------------|:---------------|---------------------|:---------------------------------------------------------------------------------------------------------|:-----------------------------------------------------------------------------------------------------------------------|
 | [**epiworldR**](https://cran.r-project.org/package=epiworldR)               | yes              | yes            | yes           | yes            | yes                 | [![status](https://tinyverse.netlify.com/badge/epiworldR)](https://CRAN.R-project.org/package=epiworldR) | [![Activity](https://img.shields.io/github/last-commit/UofUEpiBio/epiworldR)](https://github.com/UofUEpiBio/epiworldR) |
 | [**ABM**](https://cran.r-project.org/package=ABM)                           | \-               | \-             | \-            | yes            | yes                 | [![status](https://tinyverse.netlify.com/badge/ABM)](https://CRAN.R-project.org/package=ABM)             | [![Activity](https://img.shields.io/github/last-commit/junlingm/ABM)](https://github.com/junlingm/ABM)                 |
[…]
