+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: epiworldR
 Type: Package
 Title: Fast Agent-Based Epi Models
-Version: 0.0-2
-Date: 2023-06-16
+Version: 0.0-3
+Date: 2023-08-29
 Authors@R: c(
   person("Derek", "Meyer", role=c("aut","cre"), email="derekmeyer37@gmail.com"),
   person("George", "Vega Yon", role=c("aut"),
