+++ DESCRIPTION
@@ -4,9 +4,15 @@ Title: Fast Agent-Based Epi Models
 Version: 0.0-3
 Date: 2023-08-29
 Authors@R: c(
-  person("Derek", "Meyer", role=c("aut","cre"), email="derekmeyer37@gmail.com"),
+  person("Derek", "Meyer", role=c("aut","cre"), 
+  email="derekmeyer37@gmail.com", comment = c(ORCID = "0009-0005-1350-6988")),
   person("George", "Vega Yon", role=c("aut"),
   email="g.vegayon@gmail.com", comment = c(ORCID = "0000-0002-3171-0844")),
+  person("Susan", "Holmes", role = "rev", comment =
+         c(what = "JOSS reviewer", ORCID="0000-0002-2208-8168")),
+  person("Abinash", "Satapathy", role = "rev", comment =
+         c(what = "JOSS reviewer", ORCID="0000-0002-2955-2744")),
+  person("Carinogurjao", role = "rev"),
   person("Centers for Disease Control and Prevention", role="fnd", comment = "Award number 1U01CK000585; 75D30121F00003"
   ))
 Description: A flexible framework for Agent-Based Models (ABM), the 'epiworldR' package provides methods for prototyping disease outbreaks and transmission models using a 'C++' backend, making it very fast. It supports multiple epidemiological models, including the Susceptible-Infected-Susceptible (SIS), Susceptible-Infected-Removed (SIR), Susceptible-Exposed-Infected-Removed (SEIR), and others, involving arbitrary mitigation policies and multiple-disease models. Users can specify infectiousness/susceptibility rates as a function of agents' features, providing great complexity for the model dynamics. Furthermore, 'epiworldR' is ideal for simulation studies featuring large populations.
