### 2023-09-01 #5781 @Abinashbunty
@spholmes My review is complete. Following is the feedback:

- [x] Functionality: I created a fork of the repo and executed all the files, examples and tests. All executions resulted in the expected responses as documented in the repo. 🚀
- [x] Documentation: Each function and file has been referenced properly. 👍🏼
- [x] Software Paper: Clear description of features, statement of need and conclusion. 👍🏼

I can conclude that my review is complete and it is up for further processing for JOSS.
