### 2023-09-04 #5781 @spholmes
@derekmeyer37 , in order to finalize the acceptance, could you check the paper one last time (especially for the references to the CRAN packages, I am not sure they have DOIs, just check if any of those packages have associated publications that would be better references), then please make a tagged release and archive, and report the version number and archive DOI in this review thread. Thanks....
