### 2023-09-05 d89d772
Checking if nthreads = 1 to all
