+++ vignettes/run-multiple.Rmd
@@ -60,7 +60,7 @@ parallel computing.
 # Generating a saver
 saver <- make_saver("total_hist", "reproductive")
 # Running and printing
-run_multiple(model_seirconn, ndays = 50, nsims = 50, saver = saver, nthread = 2)
+run_multiple(model_seirconn, ndays = 50, nsims = 50, saver = saver, nthread = 1)
 
 ```
 
