### 2023-09-06 65510f4
epiworld.hpp replaced by full library (cc @derekmeyer37)
