+++ inst/include/epiworld/randgraph.hpp
@@ -1,0 +1,529 @@
+#ifndef EPIWORLD_RANDGRA
+#define EPIWORLD_RANDGRA
+
+template<typename TSeq>
+class Model;
+
+template<typename TSeq>
+class Agent;
+
+class AdjList;
+
+
+template<typename TSeq, typename TDat>
+inline void rewire_degseq(
+    TDat * agents,
+    Model<TSeq> * model,
+    epiworld_double proportion
+    );
+
+template<typename TSeq = int>
+inline void rewire_degseq(
+    std::vector< Agent<TSeq> > * agents,
+    Model<TSeq> * model,
+    epiworld_double proportion
+    )
+{
+
+    #ifdef EPI_DEBUG
+    std::vector< int > _degree0(agents->size(), 0);
+    for (size_t i = 0u; i < _degree0.size(); ++i)
+        _degree0[i] = model->get_agents()[i].get_neighbors().size();
+    #endif
+
+    // Identifying individuals with degree > 0
+    std::vector< epiworld_fast_uint > non_isolates;
+    std::vector< epiworld_double > weights;
+    epiworld_double nedges = 0.0;
+    
+    for (epiworld_fast_uint i = 0u; i < agents->size(); ++i)
+    {
+        if (agents->operator[](i).get_neighbors().size() > 0u)
+        {
+            non_isolates.push_back(i);
+            epiworld_double wtemp = static_cast<epiworld_double>(
+                agents->operator[](i).get_neighbors().size()
+                );
+            weights.push_back(wtemp);
+            nedges += wtemp;
+        }
+    }
+
+    if (non_isolates.size() == 0u)
+        throw std::logic_error("The graph is completely disconnected.");
+
+    // Cumulative probs
+    weights[0u] /= nedges;
+    for (epiworld_fast_uint i = 1u; i < non_isolates.size(); ++i)
+    {
+         weights[i] /= nedges;
+         weights[i] += weights[i - 1u];
+    }
+
+    // Only swap if needed
+    epiworld_fast_uint N = non_isolates.size();
+    epiworld_double prob;
+    int nrewires = floor(proportion * nedges);
+    while (nrewires-- > 0)
+    {
+
+        // Picking egos
+        prob = model->runif();
+        int id0 = N - 1;
+        for (epiworld_fast_uint i = 0u; i < N; ++i)
+            if (prob <= weights[i])
+            {
+                id0 = i;
+                break;
+            }
+
+        prob = model->runif();
+        int id1 = N - 1;
+        for (epiworld_fast_uint i = 0u; i < N; ++i)
+            if (prob <= weights[i])
+            {
+                id1 = i;
+                break;
+            }
+
+        // Correcting for under or overflow.
+        if (id1 == id0)
+            id1++;
+
+        if (id1 >= static_cast<int>(N))
+            id1 = 0;
+
+        Agent<TSeq> & p0 = agents->operator[](non_isolates[id0]);
+        Agent<TSeq> & p1 = agents->operator[](non_isolates[id1]);
+
+        // Picking alters (relative location in their lists)
+        // In this case, these are uniformly distributed within the list
+        int id01 = std::floor(p0.get_n_neighbors() * model->runif());
+        int id11 = std::floor(p1.get_n_neighbors() * model->runif());
+
+        // When rewiring, we need to flip the individuals from the other
+        // end as well, since we are dealing withi an undirected graph
+        
+        // Finding what neighbour is id0
+        model->get_agents()[id0].swap_neighbors(
+            model->get_agents()[id1],
+            id01,
+            id11
+            );
+        
+
+    }
+
+    #ifdef EPI_DEBUG
+    for (size_t _i = 0u; _i < _degree0.size(); ++_i)
+    {
[…]
