+++ src/epiworld-common.h
@@ -1,6 +1,6 @@
 #ifndef EPIWORLD_COMMON_HPP
 #define EPIWORLD_COMMON_HPP
 
-#include "epiworld.hpp"
+#include "epiworld/epiworld.hpp"
 
 #endif

[…]
