### 2023-09-06 6a1e9e3
Adding nthreads = 2 back
