+++ R/make_saver.R
@@ -41,7 +41,7 @@
 #' saver <- make_saver("total_hist", "reproductive")
 #' 
 #' # Running and printing
-#' run_multiple(model_sir, ndays = 100, nsims = 50, saver = saver, nthread = 1)
+#' run_multiple(model_sir, ndays = 100, nsims = 50, saver = saver, nthreads = 2)
 #' 
 #' # Retrieving the results
 #' ans <- run_multiple_get_results(model_sir)
