### 2023-09-06 #5781 @derekmeyer37
@spholmes , I am unfamiliar with how to "make a tagged release and archive, and report the version number and archive DOI" as this is my first paper. Could you provide me with instructions on how to do this?
