### 2023-09-06 #5781 @Abinashbunty
@derekmeyer37 You can follow this page: https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance

Moreover, for making a tagged release on GitHub, you can do something like this: https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release
