### 2023-09-07 #5781 @spholmes
🙇‍♀️ Thanks @Abinashbunty , I was just coming here with those links...you beat me to it.
@derekmeyer37 : The other step involves using zenedo or figshare to create the DOI.
Let us know if you are having trouble.
