### 2023-09-07 #5781 @derekmeyer37
@Abinashbunty @spholmes , Thanks! When uploading to Zenodo, do I need to upload all files in the package, or just the tar.gz file?
