### 2023-09-07 #5781 @spholmes
@derekmeyer37 : I think the simplest is to follow the github workflow here:https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
