### 2023-09-07 #5781 @Abinashbunty
> @derekmeyer37 : I think the simplest is to follow the github workflow here:https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

@derekmeyer37 If you follow the steps as mentioned in the link, it will be simpler on your end. 😄
![image](https://github.com/openjournals/joss-reviews/assets/25522048/8185e212-4e82-466d-b40a-a2bcaa62873e)

Zenodo will do the work for you once you toggle the button for your repository. It does automatically. Let us know in case you face some issues. 👍🏼
