### 2023-09-07 c1f9303
Adding new docker image for debugging. Fixed difference in rec rates
