+++ inst/include/epiworld/models/seird.hpp
@@ -15,14 +15,13 @@
 template<typename TSeq = int>
 class ModelSEIRD : public epiworld::Model<TSeq>
 {
-private:
+  
+public:
   static const int SUSCEPTIBLE = 0;
   static const int EXPOSED     = 1;
   static const int INFECTED    = 2;
   static const int REMOVED     = 3;
   static const int DECEASED    = 4;
-  
-public:
   
   ModelSEIRD() {};
   
@@ -66,65 +65,60 @@ public:
   ) -> void {
     
     auto state = p->get_state();
-    
-    if (state == ModelSEIRD<TSeq>::INFECTED)
+      
+    // Odd: Die, Even: Recover
+    epiworld_fast_uint n_events = 0u;
+    for (const auto & v : p->get_viruses())
     {
       
+      // Die
+      m->array_double_tmp[n_events++] = 
+        v->get_prob_death(m) * (1.0 - p->get_death_reduction(v, m)); 
       
-      // Odd: Die, Even: Recover
-      epiworld_fast_uint n_events = 0u;
-      for (const auto & v : p->get_viruses())
-      {
-        
-        // Die
-        m->array_double_tmp[n_events++] = 
-          v->get_prob_death(m) * (1.0 - p->get_death_reduction(v, m)); 
-        
-        // Recover
-        m->array_double_tmp[n_events++] = 
-          1.0 - (1.0 - v->get_prob_recovery(m)) * (1.0 - p->get_recovery_enhancer(v, m)); 
-        
-      }
+      // Recover
+      m->array_double_tmp[n_events++] = 
+        1.0 - (1.0 - v->get_prob_recovery(m)) * (1.0 - p->get_recovery_enhancer(v, m)); 
       
+    }
+    
 #ifdef EPI_DEBUG
-      if (n_events == 0u)
-      {
-        printf_epiworld(
-          "[epi-debug] agent %i has 0 possible events!!\n",
-          static_cast<int>(p->get_id())
-        );
-        throw std::logic_error("Zero events in exposed.");
-      }
+    if (n_events == 0u)
+    {
+      printf_epiworld(
+        "[epi-debug] agent %i has 0 possible events!!\n",
+        static_cast<int>(p->get_id())
+      );
+      throw std::logic_error("Zero events in exposed.");
+    }
 #else
-      if (n_events == 0u)
-        return;
+    if (n_events == 0u)
+      return;
 #endif
+    
+    
+    // Running the roulette
+    int which = roulette(n_events, m);
+    
+    if (which < 0)
+      return;
+    
+    // Which roulette happen?
+    if ((which % 2) == 0) // If odd
+    {
       
+      size_t which_v = std::ceil(which / 2);
+      p->rm_agent_by_virus(which_v, m);
       
-      // Running the roulette
-      int which = roulette(n_events, m);
+    } else {
       
-      if (which < 0)
-        return;
+      size_t which_v = std::floor(which / 2);
+      p->rm_virus(which_v, m);
       
-      // Which roulette happen?
-      if ((which % 2) == 0) // If odd
-      {
-        
-        size_t which_v = std::ceil(which / 2);
-        p->rm_agent_by_virus(which_v, m);
-        
-      } else {
-        
-        size_t which_v = std::floor(which / 2);
-        p->rm_virus(which_v, m);
-        
-      }
+    }
+    
+    return ;
       
-      return ;
-      
-    } else
-      throw std::logic_error("This function can only be applied to exposed or infected individuals. (SEIRD)") ;
+    
     
     return;
     
@@ -165,6 +159,7 @@ inline ModelSEIRD<TSeq>::ModelSEIRD(
   virus.set_prob_infecting(&model("Transmission rate"));
[…]
