+++ playground/transition_matrix_checker.Rmd
@@ -6,23 +6,25 @@ output: html_document
 ---
 
 # SEIRD Example
-```{r}
-#' model_seird <- ModelSEIRD(name = "COVID-19", prevalence = 0.01, 
-#' transmission_rate = 0.9, recovery_rate = 0.2, incubation_days = 4, 
-#' death_rate = 0.1)
-#' 
-#' # Adding a small world population
-#' agents_smallworld(
-#'   model_seird,
-#'   n = 500000,
-#'   k = 5,
-#'   d = FALSE,
-#'   p = .01
-#'   )
-#'   
-#' # Running and printing
-#' run(model_seird, ndays = 100, seed = 1912)
-#' model_seird
+```{r eval=FALSE}
+library(epiworldR)
+
+model_seird <- ModelSEIRD(name = "COVID-19", prevalence = 0.01, 
+transmission_rate = 0.9, recovery_rate = 0.2, incubation_days = 4, 
+death_rate = 0.1)
+
+# Adding a small world population
+agents_smallworld(
+  model_seird,
+  n = 500000,
+  k = 5,
+  d = FALSE,
+  p = .01
+  )
+  
+# Running and printing
+run(model_seird, ndays = 100, seed = 1912)
+model_seird
 
 # Transition Probabilities:
 #  - Susceptible  0.96  0.04  0.00  0.00  0.00
@@ -33,36 +35,38 @@ output: html_document
 
 pdie <- 0.1
 precovery <- 0.2
-pnone <- (1-pdie)*(1-precovery)
-patmost_one <- pdie*(1-precovery) + precovery*(1-pdie)
+pnone <- (1 - pdie) * (1 - precovery)
+patmost_one <- pdie * (1 - precovery) + precovery * (1 - pdie)
 
 p_die_given_infected <- pdie * (1-precovery) / (pnone + patmost_one) # 0.08163265
-p_recovery_given_infected <- precovery * (1-pdie) / (pnone + patmost_one) # 0.1836735
+p_recovery_given_infected <- precovery * (1 - pdie) / (pnone + patmost_one) # 0.1836735
 
+p_neither <- (1 - pdie) * (1 - precovery) / (pnone + patmost_one) # 0.7346939
 ```
 
 $P(Dying | Infected) = \frac{p_{die} * (1-p_{recovery})}{(1-p_{die})*(1-p_{recovery})+p_{die}*(1-p_{recovery}) + p_{recovery}*(1-p_{die})} = \frac{0.01 * (1-0.1)}{(1-0.01)*(1-0.1) + 0.01*(1-0.01)+0.1*(1-0.01)} = 0.009 \approx 0.01$
 
 # SEIRDCONN Example
-```{r}
-#' # An example with COVID-19
-#' model_seirdconn <- ModelSEIRDCONN(
-#'   name                = "COVID-19",
-#'   prevalence          = 0.01, 
-#'   n                   = 100000,
-#'   contact_rate        = 2, 
-#'   incubation_days     = 7, 
-#'   transmission_rate   = 0.5,
-#'   recovery_rate       = 0.3,
-#'   death_rate          = 0.01
-#' )
-#'   
-#' # Running and printing
-#' run(model_seirdconn, ndays = 100, seed = 1912)
-#' model_seirdconn
-#' 
-#' plot(model_seirdconn)
+```{r eval = FALSE}
+# An example with COVID-19
+model_seirdconn <- ModelSEIRDCONN(
+  name                = "COVID-19",
+  prevalence          = 0.01, 
+  n                   = 100000,
+  contact_rate        = 2, 
+  incubation_days     = 7, 
+  transmission_rate   = 0.5,
+  recovery_rate       = 0.3,
+  death_rate          = 0.01
+)
+  
+# Running and printing
[…]
