+++ docker/Makefile
@@ -1,0 +1,2 @@
+all:
+	docker build -t uofuepibio/epiworldr:debug -f Dockerfile .

[…]
