### 2023-09-07 #5781 @derekmeyer37
@spholmes , here is that information for you. Thank you for your help and please let me know if I can provide further information.
Version Number: v0.3-0
DOI: https://doi.org/10.5281/zenodo.8327345
