### 2023-09-09 #5781 @spholmes
@derekmeyer37 and @Abinashbunty : There is a problem of inconsistency between the package authors and the archive authors, for the tagged release we need the same authors as on the JOSS paper that was submitted and the current version of the software, the reviewer should not appear as a co-author of the archive as is currently the case.
Please see on this github issue how to change the meta-data on zenodo as detailed in the following issue:
https://github.com/openjournals/joss/pull/1040
