### 2023-09-15 #5781 @spholmes
@derekmeyer37 : Just a tiny adjustment to the references, could you put the R in curly brackets {R}
so that the title of the reference Gochanour, B. et al. is:
abmR: Agent-based models in R. instead of a lower case r..
Thanks, I think we are almost there.
