### 2023-09-20 7db6fd0
Update paper.bib [skip ci]
