+++ paper.bib
@@ -7,7 +7,7 @@
 }
 
 @Manual{abmR,
-    title = {abmR: Agent-Based Models in R},
+    title = {abmR: Agent-Based Models in {R}},
     author = {Benjamin Gochanour and Javier Fernandez-Lopez and Andrea Contina},
     year = {2023},
     note = {R package version 1.0.9},
