### 2023-09-20 #5781 @spholmes
@derekmeyer37 I think the DOI problem is non essential since these are cran packages for which the link works. I am just going to accept ...
