### 2023-09-26 #5781 @Kevin-Mattheus-Moerman
@spholmes thanks for editing here. I have some tips for the future:

* Please have the version tag fully match the release tag on GitHub, this include the `v` here so I updated the version tag for JOSS.
* Note you can now call `@editorialbot create post-review checklist` (see also: https://joss.readthedocs.io/en/latest/editorial_bot.html#post-review-checklist), this is often helpful to make sure no final steps/checks are missed (such as the license/title on the archive link matching etc).
* The above DOIs should still be resolved, I'll ask the authors to do this now.
