### 2023-09-26 #5781 @Kevin-Mattheus-Moerman
@derekmeyer37 I am the AEiC for this track and here to help process this paper for acceptance in JOSS. I have checked the software repository, the archive link, this review, and the paper. Below are some final points that require your attention:

- [ ] Please use the DOI (I think a full URL is give), so just `10.1111/j.2041-210x.2011.00180.x` for the following reference:
```
Thiele, J. C., Kurth, W., & Grimm, V. (2012). RNetLogo: An R package for running and
exploring individual-based models implemented in NetLogo. Methods in Ecology and
Evolution, 3(3), 480–483. http://onlinelibrary.wiley.com/doi/10.1111/j.2041-210X.2011.
00180.x/abstract
```

- [ ]  Please use the following DOI `10.1111/2041-210X.14014` for the this reference (note our editorialbot picked up on a pre-print DOI, however that helps point out this accepted version DOI which is what we should use):
```
Ma, J. (2023). ABM: Agent based model simulation framework. https://CRAN.R-project.
org/package=ABM
```
Note :point_up: I am assuming it is appropriate to cite this paper for this package? (most authors would prefer if papers rather than CRAN links receive citations). If instead you want to cite the CRAN link I suggest citing both to give these authors the right "academic credit".

- [ ]  Please update the ZENODO listed license, it currently says `Other (Open)`, to match your software license (MIT license).
- [ ]  Please update the ZENODO listed title (currently `UofUEpiBio/epiworldR: v0.3-0`) to match your paper title: `epiworldR: Fast Agent-Based Epi Models`.
- [ ]  Please update the ZENODO listed author order to match that of the paper.
- [ ]  Please remove the `Editors` information listed for your ZENODO archive link.
- [ ] Recommendation: Currently GitHub lists the following for your license: `Unknown, MIT licenses found`. This is because your plain text `LICENSE` file only contains a copyright statement. Then you also have a `LICENSE.md` file which contains both a copyright statement and the MIT License description. I recommend that you remove the `LICENSE.md` file, and instead use only the `LICENSE` file to show the copyright and MIT license text. This will help GitHub properly detect your license. It may be that CRAN has a funny license format, which is why you've chosen your approach. If so, so be it. As a second recommendation though I would at the very least add a License heading to the README and link to the license file.
