### 2023-09-27 5bdeb7c
Update paper.bib [skip ci]
