+++ paper.bib
@@ -11,7 +11,7 @@
     author = {Benjamin Gochanour and Javier Fernandez-Lopez and Andrea Contina},
     year = {2023},
     note = {R package version 1.0.9},
-    url = {https://CRAN.R-project.org/package=abmR},
+    url = {10.1111/2041-210X.14014},
 }
   
 @article{gochanour2021abmr,
@@ -49,5 +49,5 @@
     number = {3},
     pages = {480--483},
     url =
-      {http://onlinelibrary.wiley.com/doi/10.1111/j.2041-210X.2011.00180.x/abstract},
+      {10.1111/j.2041-210X.2011.00180.x},
 }
