### 2023-10-03 b3d446e
Update paper.bib
Fixed doi entries in paper.bib
