+++ paper.bib
@@ -11,16 +11,17 @@
     author = {Benjamin Gochanour and Javier Fernandez-Lopez and Andrea Contina},
     year = {2023},
     note = {R package version 1.0.9},
-    url = {10.1111/2041-210X.14014},
+    doi = {10.1111/2041-210X.14014},
 }
   
 @article{gochanour2021abmr,
-  title={abmR: An R package for agent-based model analysis of large-scale movements across taxa},
-  author={Gochanour, Benjamin and Fern{\'a}ndez-L{\'o}pez, Javier and Contina, Andrea},
-  journal={bioRxiv},
-  pages={2021--09},
-  year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  title = {abmR: An R package for agent-based model analysis of large-scale movements across taxa},
+  author = {Gochanour, Benjamin and Fern{\'a}ndez-L{\'o}pez, Javier and Contina, Andrea},
+  journal = {bioRxiv},
+  pages = {2021--09},
+  year = {2021},
+  publisher = {Cold Spring Harbor Laboratory},
+  doi = {10.1101/2021.09.15.460374}
 }
 
 @Manual{cystiSim,
@@ -48,6 +49,5 @@
     volume = {3},
     number = {3},
     pages = {480--483},
-    url =
-      {10.1111/j.2041-210X.2011.00180.x},
+    doi = {10.1111/j.2041-210X.2011.00180.x},
 }
