### 2023-10-03 #5781 @Kevin-Mattheus-Moerman
@derekmeyer37 Your `.bib` file currently uses the `url` field e.g. like this:
```
url = {10.1111/2041-210X.14014}
```
However, this should be the `doi` field, e.g. like this:
```
doi = {10.1111/2041-210X.14014}
```
I've tried to address this in [this pull request](https://github.com/UofUEpiBio/epiworldR/pull/8), but I haven't tested paper compilation yet so do check first.
