### 2023-10-03 bfa78c2
Update paper.bib [skip ci]
