+++ paper.bib
@@ -11,7 +11,7 @@
     author = {Benjamin Gochanour and Javier Fernandez-Lopez and Andrea Contina},
     year = {2023},
     note = {R package version 1.0.9},
-    url = {10.1111/2041-210X.14014},
+    doi = {10.1111/2041-210X.14014},
 }
   
 @article{gochanour2021abmr,
@@ -48,6 +48,5 @@
     volume = {3},
     number = {3},
     pages = {480--483},
-    url =
-      {10.1111/j.2041-210X.2011.00180.x},
+    doi = {10.1111/j.2041-210X.2011.00180.x},
 }
