### 2023-10-04 #5781 @Kevin-Mattheus-Moerman
@derekmeyer37 congratulations on this publication in JOSS!

Thanks for editing @spholmes, and a special thanks to the reviewers: @Abinashbunty, @carinogurjao
